convert to long

+++ features/logistic/logisticcart/src/main/java/com/tokopedia/logisticcart/shipping/model/ShipmentCartItemModel.java
@@ -76,7 +76,7 @@ public class ShipmentCartItemModel implements Parcelable {
 
     private boolean isFulfillment;
     private String fulfillmentBadgeUrl;
-    private int fulfillmentId;
+    private long fulfillmentId;
 
     // promo stacking
     private boolean hasPromoList;
@@ -145,7 +145,7 @@ public class ShipmentCartItemModel implements Parcelable {
         blackboxInfo = in.readString();
         isFulfillment = in.readByte() != 0;
         fulfillmentBadgeUrl = in.readString();
-        fulfillmentId = in.readInt();
+        fulfillmentId = in.readLong();
         hasPromoList = in.readByte() != 0;
         voucherLogisticItemUiModel = in.readParcelable(VoucherLogisticItemUiModel.class.getClassLoader());
         isLeasingProduct = in.readByte() != 0;
@@ -201,7 +201,7 @@ public class ShipmentCartItemModel implements Parcelable {
         dest.writeString(blackboxInfo);
         dest.writeByte((byte) (isFulfillment ? 1 : 0));
         dest.writeString(fulfillmentBadgeUrl);
-        dest.writeInt(fulfillmentId);
+        dest.writeLong(fulfillmentId);
         dest.writeByte((byte) (hasPromoList ? 1 : 0));
         dest.writeParcelable(voucherLogisticItemUiModel, flags);
         dest.writeByte((byte) (isLeasingProduct ? 1 : 0));
@@ -562,11 +562,11 @@ public class ShipmentCartItemModel implements Parcelable {
         this.fulfillmentBadgeUrl = fulfillmentBadgeUrl;
     }
 
-    public int getFulfillmentId() {
+    public long getFulfillmentId() {
         return fulfillmentId;
     }
 
-    public void setFulfillmentId(int fulfillmentId) {
+    public void setFulfillmentId(long fulfillmentId) {
         this.fulfillmentId = fulfillmentId;
     }
