+++ features/transaction/checkout/src/main/java/com/tokopedia/checkout/data/model/response/shipmentaddressform/Warehouse.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Warehouse(
         @SuppressLint("Invalid Data Type")
         @SerializedName("warehouse_id")
-        val warehouseId: Int = 0,
+        val warehouseId: Long = 0,
         @SerializedName("city_name")
         val cityName: String = ""
 )
